double based implementation

+++ src/main/java/com/serdarmumcu/addtwonumbers/bean/AddRequest.java
@@ -2,27 +2,27 @@ package com.serdarmumcu.addtwonumbers.bean;
 
 public class AddRequest {
 
-    private int num1;
-    private int num2;
+    private double num1;
+    private double num2;
 
-    public AddRequest(int num1, int num2) {
+    public AddRequest(double num1, double num2) {
         this.num1 = num1;
         this.num2 = num2;
     }
 
-    public int getNum1() {
+    public double getNum1() {
         return num1;
     }
 
-    public void setNum1(int num1) {
+    public void setNum1(double num1) {
         this.num1 = num1;
     }
 
-    public int getNum2() {
+    public double getNum2() {
         return num2;
     }
 
-    public void setNum2(int num2) {
+    public void setNum2(double num2) {
         this.num2 = num2;
     }
 }
